Cleaned up (some of) the mess in ParseGL
ScanParams now stores its results in a ScanParamsResult object, instead of the global arrays it was using.
  I still pass a ststic instance of this object (since we're still singlethreaded) but it is at least
  better contained, and easier to maintain.
Got rid of the "auto-create" scan types, in favor of scanning the name, checking for null, and allocating if needed.
Added '*' scan type, for repeated element-type scans.  This gets rid of those stupid parameter string constants.
Moved the filename-with-spaces kluge into ScanParams, where it is more maintainable.  Still a kluge though.