Change design of default Frustum to give less clipping
Previously near clipping plane was at 1, and the origin
was *behind* it.  The modelspace was forced to translate
to something sensible.  Now, the projection matrix gets
a default translation so the clipping plane is at -19,
and at the origin, (-10,10) is the height of the Y axis
(or the width of the X axis depending on window size)