New mechanism for automatic version number, based on git
If git is not available, it uses the version from the change log.
If git is available, it uses the current tag if the HEAD is tagged
and the repository is clean.  Else it uses the tag as the version
and the commit as the suffix, also mentioning whether the work
dir was clean at the time of the build.

Also moved the change log to standard name Changes

Also removed the makefile code that would install symlinks per
OpenGL command, since the new CmdlineGL.lib is a much better design.