Replace cglFixedPt with optional divisor and push/pop default
cglFixedPt was sort of a pain to work with since the main script
might choose a value inconvenient for a library, and then the
library code needed to know what value to restore afterward.

With this new design, a divisor can always be specified without
worrying about the current divisor, and the push/pop system allows
for easy usage of the feature without regard for the caller.