Fixed handling for screen resize, and added/enhanced events
Previously the screen resize would automatically trigger a glViewport
and projection matrix recalculation.  This wasn't valid if performed
in the middle of the user rendering a frame.

Now, there is a global flag that indicates whether the frame is in
progress, and a resize event will be postponed until the current frame
is complete.

Also, the user can now control whether CmdlineGL automatically runs
glViewport and sets up the projection matrix.  This setting affects
both initial startup and after a resize.

While I was at it, I added fields to the key event to describe which
character was generated, and added some "W" events for window-related
things.