And massive rewrite AGAIN
This design keeps the best features of the old version (with the
parsing system packing the values into useful vectors that can be
passed direct to OpenGL API) and the best of the previous version
(where a Parse function breaks the input line as needed according
to a specification).

Also, instead of storing the data directly in the hash tables, make
the table out of 16-bit ints which point into a packed array of
records.  This should take up much less memory.

It shoud now be reasonably fast, compact, less code, better handling
of object creation, less buggy, etc.